improve ui

+++ src/main/java/org/strasa/web/maintenance/view/model/EditUploadedStudies.java
@@ -8,6 +8,7 @@ import java.util.Map;
 import org.strasa.middleware.manager.ProgramManagerImpl;
 import org.strasa.middleware.manager.ProjectManagerImpl;
 import org.strasa.middleware.manager.StudyManagerImpl;
+import org.strasa.middleware.manager.UserManagerImpl;
 import org.strasa.middleware.model.Study;
 import org.zkoss.bind.annotation.AfterCompose;
 import org.zkoss.bind.annotation.BindingParam;
@@ -20,6 +21,7 @@ import org.zkoss.zhtml.Messagebox;
 import org.zkoss.zk.ui.Component;
 import org.zkoss.zk.ui.Components;
 import org.zkoss.zk.ui.Executions;
+import org.zkoss.zk.ui.event.Event;
 import org.zkoss.zk.ui.select.Selectors;
 import org.zkoss.zk.ui.select.annotation.Wire;
 import org.zkoss.zul.Div;
@@ -129,11 +131,22 @@ public class EditUploadedStudies {
 		 Components.removeAllChildren(list.get(0));
 		Executions.createComponents("/user/uploadstudy/index.zul" , list.get(0), arg);
 	}
+	@SuppressWarnings("unchecked")
 	@NotifyChange("editStudyList")
 	@Command("deleteStudy")
-	public void deleteStudy(@BindingParam("studyId") Integer studyId){
-		studyMan.deleteStudyById(studyId);
-		populateEditStudyList();
+	public void deleteStudy(@BindingParam("studyId") final Integer studyId){
+		
+		
+		Messagebox.show("Are you sure you want to delete this study? This cannot be undone", "Confirm Dialog", Messagebox.OK | Messagebox.CANCEL, Messagebox.QUESTION, new org.zkoss.zk.ui.event.EventListener() {
+			public void onEvent(Event evt) throws InterruptedException {
+				if (evt.getName().equals("onOK")) {
+					studyMan.deleteStudyById(studyId);
+					populateEditStudyList();
+				} 
+			}
+		});
+		
+		
 	}
 
 	@NotifyChange("*")
